Database tényleges összekötése backendel

+++ src/main/java/hu/csongor/demo/repository/JelentkezesRepository.java
@@ -4,12 +4,12 @@ import hu.csongor.demo.entity.Jelentkezes;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 public interface JelentkezesRepository
-        extends JpaRepository<Jelentkezes, Long> {
+        extends JpaRepository<Jelentkezes, Integer> {
     boolean existsByUserIdAndSzakkorId(
-            Long userId,
-            Long szakkorId
+            Integer userId,
+            Integer szakkorId
     );
-    List<Jelentkezes> findByUserId(Long userId);
-    List<Jelentkezes> findBySzakkorId(Long szakkorId);
-    long countBySzakkorId(Long szakkorId);
+    List<Jelentkezes> findByUserId(Integer userId);
+    List<Jelentkezes> findBySzakkorId(Integer szakkorId);
+    long countBySzakkorId(Integer szakkorId);
 }
